test: post 검증 메서드 클래스로 분리
PostTest -> AssertPost
