+++ src/test/java/com/study/board/domain/post/domain/PostTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
+import static com.study.board.domain.post.AssertPost.assertPostCreation;
+import static com.study.board.domain.post.AssertPost.assertWriter;
 import static com.study.board.domain.post.domain.Post.POST_TITLE_MAX_LENGTH;
 import static com.study.board.fixture.Fixture.createUser;
 import static org.assertj.core.api.Assertions.*;
@@ -47,30 +49,5 @@ public class PostTest {
     void 작성자가_null_이면_생성_실패() {
         assertThatNullPointerException()
                 .isThrownBy(() -> Post.create("제목", "내용", null));
-    }
-
-    private void assertPostCreation(Post post) {
-        assertThat(post.getId()).isNull();
-        assertThat(post.getTitle()).isEqualTo("제목");
-        assertThat(post.getContent()).isEqualTo("내용");
-        assertThat(post.getWrittenDateTime()).isNotNull();
-    }
-
-    public static void assertPost(Post post, String title, String content) {
-        assertThat(post.getId()).isNotNull();
-        assertThat(post.getTitle()).isEqualTo(title);
-        assertThat(post.getContent()).isEqualTo(content);
-        assertThat(post.getWrittenDateTime()).isNotNull();
-    }
-
-    public static void assertWriter(User writer, Long expectedId) {
-        assertThat(writer.getId()).isEqualTo(expectedId);
-        assertThat(writer.getName()).isEqualTo("득윤");
-        assertThat(writer.getHobby()).isEqualTo("체스");
-    }
-
-    public static void assertPostWithWriter(Post post, String title, String content, Long expectedWriterId) {
-        assertPost(post, title, content);
-        assertWriter(post.getWriter(), expectedWriterId);
     }
 }
